Clean up project grid
